move utils to a sub module thing
