+++ bin/dev.js
@@ -1,5 +1,5 @@
 const nodemon = require('nodemon')
-const { Utils: { logger } } = require('../')
+const { logger } = require('../util')
 
 nodemon({
   script: 'demo/index.js',
