+++ events/disconnect.js
@@ -1,6 +1,7 @@
-const { DiscordEvent, Utils } = require('../lib')
+const { DiscordEvent } = require('../lib')
+const { logger } = require('../util')
 
 module.exports = new DiscordEvent({
   name: 'disconnect',
-  run: (bot) => Utils.logger.error('disconnected')
+  run: (bot) => logger.error('disconnected')
 })
