+++ index.js
@@ -1,7 +1,1 @@
-const Library = require('./lib')
-const PermissionLibrary = require('./permissions')
-
-module.exports = {
-  ...Library,
-  PermissionLibrary
-}
+module.exports = require('./lib')
