+++ lib/command/index.js
@@ -1,6 +1,6 @@
 const {
   ExtendedMap
-} = require('../utils')
+} = require('../../util')
 
 class Command {
   /**
