+++ lib/data-client/index.js
@@ -7,7 +7,7 @@ const {
 const Orator = require('../orator')
 const RAMManager = require('../ram-manager')
 const StatusManager = require('../status-manager')
-const { logger, ExtendedMap } = require('../utils')
+const { logger, ExtendedMap } = require('../../util')
 
 const defaults = require('../../config/settings.json')
