+++ lib/data-client/index.test.js
@@ -7,7 +7,7 @@ import {
 } from 'eris'
 
 import DataClient from '.'
-import { ExtendedMap } from '../utils'
+import { ExtendedMap } from '../../util'
 import esmCommand from './test-files/esm'
 
 const fakeDatabaseManager = {
