+++ lib/index.js
@@ -9,6 +9,5 @@ module.exports = {
   Permission: require('./permission'),
   RAMManager: require('./ram-manager'),
   SQLManager: require('./sql-manager'),
-  StatusManager: require('./status-manager'),
-  Utils: require('./utils')
+  StatusManager: require('./status-manager')
 }
