+++ lib/orator/index.js
@@ -1,4 +1,4 @@
-const { logger, get } = require('../utils')
+const { logger, get } = require('../../util')
 
 class Orator {
   /**
