+++ lib/status-manager/index.js
@@ -5,7 +5,7 @@ const {
     equalStatuses
   },
   logger
-} = require('../utils')
+} = require('../../util')
 
 class StatusManager {
   /**
